% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axel.R
\name{axel}
\alias{axel}
\title{axel downloader}
\usage{
axel(
  input_url,
  output_path,
  background = FALSE,
  nThread = 1,
  force_overwrite = FALSE,
  quiet = TRUE,
  alternate = TRUE,
  check_certificates = FALSE
)
}
\arguments{
\item{input_url}{input_url.}

\item{output_path}{output_path.}

\item{background}{Run in background}

\item{nThread}{Number of threads to parallelize over.}

\item{force_overwrite}{Overwrite existing file.}

\item{quiet}{Run quietly.}

\item{alternate}{alternate,}

\item{check_certificates}{check_certificates}
}
\value{
Path where the file has been downloaded
}
\description{
R wrapper for axel, which enables multi-threaded download
of a single large file.
}
\seealso{
\url{https://github.com/axel-download-accelerator/axel/}

Other downloaders: 
\code{\link{downloader}()}
}
\concept{downloaders}
\keyword{internal}
