% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieugwasr.R
\name{check_access_token}
\alias{check_access_token}
\title{Check if authentication has been made}
\usage{
check_access_token()
}
\value{
NULL or access_token depending on current authentication state
}
\description{
If a call to get_access_token() has been made then it will
have generated mrbase.oauth. Pass the token if it is present,
if not, return NULL and do not authenticate.
}
\keyword{internal}
