% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_bi_allelic.R
\name{check_bi_allelic}
\alias{check_bi_allelic}
\title{Remove non-biallelic SNPs}
\usage{
check_bi_allelic(sumstats_dt, path, ref_genome, bi_allelic_filter, rsids)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}

\item{ref_genome}{name of the reference genome used for the GWAS (GRCh37 or GRCh38). Default is GRCh37.}

\item{bi_allelic_filter}{Binary Should non-biallelic SNPs be removed. Default is TRUE}

\item{rsids}{datatable of snpsById, filtered to SNPs of interest if loaded already. Or else NULL}
}
\value{
A list containing two data tables:
\itemize{
\item \code{sumstats_dt}: the modified summary statistics data table object
\item \code{rsids}: snpsById, filtered to SNPs of interest if loaded already. Or else NULL
}
}
\description{
Remove non-biallelic SNPs
}
\keyword{internal}
