% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_order.R
\name{check_col_order}
\alias{check_col_order}
\title{Ensure that the first three columns are SNP, CHR, BP in that order and
then A1, A2 if present}
\usage{
check_col_order(sumstats_dt, path)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
list containing sumstats_dt, the modified summary statistics
data table object
}
\description{
Ensure that the first three columns are SNP, CHR, BP in that order and
then A1, A2 if present
}
\keyword{internal}
