% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dup_bp.R
\name{check_dup_bp}
\alias{check_dup_bp}
\title{Ensure all rows have unique positions, drop those that don't}
\usage{
check_dup_bp(sumstats_file, path)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
null
}
\description{
Ensure all rows have unique positions, drop those that don't
}
