% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dup_col.R
\name{check_dup_col}
\alias{check_dup_col}
\title{Ensure that no columns are duplicated}
\usage{
check_dup_col(sumstats_file, path)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
The modified sumstats_file
}
\description{
Ensure that no columns are duplicated
}
