% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dup_snp.R
\name{check_dup_snp}
\alias{check_dup_snp}
\title{Ensure all rows have unique SNP IDs, drop those that don't}
\usage{
check_dup_snp(sumstats_file, path)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
null
}
\description{
Ensure all rows have unique SNP IDs, drop those that don't
}
