% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_effect_columns_nonzero.R
\name{check_effect_columns_nonzero}
\alias{check_effect_columns_nonzero}
\title{Ensure that the standard error (se) is positive for all SNPs}
\usage{
check_effect_columns_nonzero(sumstats_dt, path, effect_columns_nonzero)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted}

\item{effect_columns_nonzero}{Binary should the effect columns in the data
BETA,OR (odds ratio),LOG_ODDS,SIGNED_SUMSTAT be checked to ensure no SNP=0.
Those that do are removed(if present in sumstats file). Default TRUE.}
}
\value{
list containing sumstats_dt, the modified summary statistics data
table object
}
\description{
Ensure that the standard error (se) is positive for all SNPs
}
\keyword{internal}
