% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empty_cols.R
\name{check_empty_cols}
\alias{check_empty_cols}
\title{Check for empty columns}
\usage{
check_empty_cols(sumstats_dt, sampled_rows = NULL, verbose = TRUE)
}
\arguments{
\item{sampled_rows}{First N rows to sample.
Set \code{NULL} to use full \code{sumstats_file}.
when determining whether cols are empty.}

\item{verbose}{Print messages.}
}
\value{
empty_cols
}
\description{
Empty columns contain only ".", NA, or 0
}
\keyword{internal}
