% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_info_score.R
\name{check_info_score}
\alias{check_info_score}
\title{Ensure all SNPs have info score above threshold}
\usage{
check_info_score(sumstats_dt, path, INFO_filter)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}

\item{INFO_filter}{integer The minimum value permissible of the imputation information score (if present in sumstatsfile). Default 0.9}
}
\value{
list containing sumstats_dt, the modified summary statistics data table object
}
\description{
Ensure all SNPs have info score above threshold
}
\keyword{internal}
