% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ldsc_format.R
\name{check_ldsc_format}
\alias{check_ldsc_format}
\title{Ensures that parameters are compatible with LDSC format}
\source{
\href{https://github.com/bulik/ldsc}{LDSC GitHub}
}
\usage{
check_ldsc_format(ldsc_format, convert_n_int, allele_flip_check, compute_z)
}
\arguments{
\item{ldsc_format}{Ensure that output format meets all requirements
to be fed directly into LDSC without the need for additional munging.}

\item{convert_n_int}{Binary, if N (the number of samples) is not an integer, should this be rounded? Default is TRUE.}

\item{allele_flip_check}{Binary Should the allele columns be checked against reference genome to infer if flipping is necessary. Default is TRUE.}

\item{compute_z}{Whether to compute Z-score column from P.}
}
\value{
Formatted summary statistics
}
\description{
Format summary statistics for direct input to
Linkage Disequilibrium SCore (LDSC) regression without the need
to use their \code{munge_sumstats.py} script first.
}
\details{
\href{https://github.com/bulik/ldsc/wiki/Summary-Statistics-File-Format}{LDSC documentation}.
}
