% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_miss_data.R
\name{check_miss_data}
\alias{check_miss_data}
\title{Ensure that only one model in GWAS sumstats or only one trait tested}
\usage{
check_miss_data(sumstats_file, path)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
The modified sumstats_file
}
\description{
Ensure that only one model in GWAS sumstats or only one trait tested
}
