% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_int.R
\name{check_n_int}
\alias{check_n_int}
\title{Ensure that the N column is all integers}
\usage{
check_n_int(sumstats_dt, path, convert_n_int, imputation_ind)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}

\item{convert_n_int}{Binary, if N (the number of samples) is not an integer,
should this be rounded? Default is TRUE.}

\item{imputation_ind}{Binary Should a column be added for each imputation
step to show what SNPs have imputed values for differing fields. This
includes a field denoting SNP allele flipping (flipped). \strong{Note}
these columns will be in the formatted summary statistics returned. Default
is FALSE.}
}
\value{
list containing sumstats_dt, the modified summary
statistics data table object.
}
\description{
Ensure that the N column is all integers
}
\keyword{internal}
