% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_no_allele.R
\name{check_no_allele}
\alias{check_no_allele}
\title{Ensure that A1 & A2 are present, if not can find it with SNP and other allele}
\usage{
check_no_allele(sumstats_file, path, ref_genome)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}

\item{ref_genome}{name of the reference genome used for the GWAS (GRCh37 or GRCh38)}
}
\value{
The modified sumstats_file
}
\description{
Ensure that A1 & A2 are present, if not can find it with SNP and other allele
}
