% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_no_chr_bp.R
\name{check_no_chr_bp}
\alias{check_no_chr_bp}
\title{Ensure that CHR and BP are missing if SNP is present, can find them}
\usage{
check_no_chr_bp(sumstats_file)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}
}
\value{
The modified sumstats_file
}
\description{
Ensure that CHR and BP are missing if SNP is present, can find them
}
