% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_numeric.R
\name{check_numeric}
\alias{check_numeric}
\title{Check numeric columns}
\usage{
check_numeric(sumstats_dt, cols = c("P", "SE", "FRQ", "MAF", "BETA"))
}
\arguments{
\item{sumstats_dt}{Summary stats with column names already standardised by
\link[MungeSumstats]{format_sumstats}.}

\item{cols}{Names of columns that should be numeric.
If any of these columns are not actually present in \code{sumstats_dt},
they will be skipped.}
}
\value{
sumstats_dt
}
\description{
Checks for any columns that should be numeric,
and ensures that they are indeed numeric.
}
\keyword{internal}
