% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pos_se.R
\name{check_pos_se}
\alias{check_pos_se}
\title{Ensure that the standard error (se) is positive for all SNPs}
\usage{
check_pos_se(sumstats_dt, path, pos_se)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted}

\item{pos_se}{Binary Should the standard Error (SE) column be checked to
ensure it is greater than 0? Those that are, are removed (if present in
sumstats file). Default TRUE.}
}
\value{
list containing sumstats_dt, the modified summary statistics data
table object
}
\description{
Ensure that the standard error (se) is positive for all SNPs
}
\keyword{internal}
