% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_row_snp.R
\name{check_row_snp}
\alias{check_row_snp}
\title{Ensure all rows have SNPs beginning with rs or SNP, drop those that don't}
\usage{
check_row_snp(sumstats_file, path)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
The modified sumstats_file
}
\description{
Ensure all rows have SNPs beginning with rs or SNP, drop those that don't
}
