% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tabular.R
\name{check_tabular}
\alias{check_tabular}
\title{Ensure valid tabular format}
\usage{
check_tabular(header)
}
\arguments{
\item{header}{The summary statistics file for the GWAS}
}
\value{
Whether the file is tabular
}
\description{
Ensure valid tabular format
}
\keyword{internal}
