% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_two_step_col.R
\name{check_two_step_col}
\alias{check_two_step_col}
\title{Ensure that CHR:BP aren't merged into 1 column}
\usage{
check_two_step_col(sumstats_file)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}
}
\value{
The modified sumstats_file
}
\description{
Ensure that CHR:BP aren't merged into 1 column
}
