% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vcf.R
\name{check_vcf}
\alias{check_vcf}
\title{Check if the inputted file is in VCF format}
\usage{
check_vcf(header)
}
\arguments{
\item{header}{Header of the GWAS summary statistics file.}
}
\value{
Whether the file is vcf or not
}
\description{
Check if the inputted file is in VCF format
}
\keyword{internal}
