% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vital_col.R
\name{check_vital_col}
\alias{check_vital_col}
\title{Ensure that all necessary columns are in the summary statistics file}
\usage{
check_vital_col(sumstats_file)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}
}
\value{
null
}
\description{
Ensure that all necessary columns are in the summary statistics file
}
