% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nsize.R
\name{compute_nsize}
\alias{compute_nsize}
\title{Check for N column if not present and user wants, impute N based on user's
sample size. \strong{NOTE} this will be the same value for each SNP which is not
necessarily correct and may cause issues down the line.}
\usage{
compute_nsize(
  sumstats_dt,
  imputation_ind = FALSE,
  compute_n = c("ldsc", "sum"),
  force_new = FALSE
)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the
GWAS.}

\item{imputation_ind}{Binary Should a column be added for each imputation
step to show what SNPs have imputed values for differing fields. This
includes a field denoting SNP allele flipping (flipped). \strong{Note}
these columns will be in the formatted summary statistics returned. Default
is FALSE.}

\item{compute_n}{How to compute per-SNP sample size (new column "N").
\itemize{
\item{\code{0}: }{N will not be computed.}
\item{\code{>0}: }{If any number >0 is provided,
that value will be set as N for every row.
\strong{Note}: Computing N this way is incorrect and should be avoided
if at all possible.}
\item{\code{"sum"}: }{N will be computed as:
cases (N_CAS) + controls (N_CON), so long as both columns are present}.
\item{\code{"effective"}: }{N will be computed as effective sample size:
cases (N_CAS) + controls (N_CON), so long as both columns are present}.
}}
}
\value{
\code{list("sumstats_dt"=sumstats_dt)}
}
\description{
Check for N column if not present and user wants, impute N based on user's
sample size. \strong{NOTE} this will be the same value for each SNP which is not
necessarily correct and may cause issues down the line.
}
\keyword{internal}
