% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sumstats.R
\name{convert_sumstats}
\alias{convert_sumstats}
\title{Convert sumstats to desired object type}
\usage{
convert_sumstats(
  sumstats_dt,
  return_format = c("data.table", "vranges", "granges")
)
}
\arguments{
\item{return_format}{Object type to convert to}
}
\description{
Convert sumstats to desired object type
}
\keyword{internal}
