% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF_to_dt.R
\name{DF_to_dt}
\alias{DF_to_dt}
\title{DataFrame to data.table}
\source{
\href{https://support.bioconductor.org/p/66874/}{
Solution from Bioc forum}
}
\usage{
DF_to_dt(DF)
}
\arguments{
\item{DF}{\link[S4Vectors]{DataFrame} object.}
}
\description{
Efficiently convert \link[S4Vectors]{DataFrame} to
\link[data.table]{data.table}.
}
\keyword{internal}
