% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vcf.R
\name{download_vcf}
\alias{download_vcf}
\title{Download VCF file and its index file from Open GWAS}
\usage{
download_vcf(
  vcf_url,
  vcf_dir,
  vcf_download = TRUE,
  download_method = "download.file",
  force_new = FALSE,
  quiet = TRUE,
  nThread = 1
)
}
\arguments{
\item{vcf_dir}{Where to download the original VCF from Open GWAS.}

\item{vcf_download}{Download the original VCF from Open GWAS.}

\item{download_method}{\code{"axel"} (multi-threaded) or \code{"download.file"} (single-threaded) .}

\item{force_new}{Overwrite a previously downloaded VCF with the same path name.}

\item{quiet}{Run quietly.}

\item{nThread}{Number of threads to parallelize over.}
}
\description{
Ideally, we would use \href{https://github.com/MRCIEU/gwasvcf}{gwasvcf} instead
but it hasn't been made available on CRAN or Bioconductor yet,
so we can't include it as a dep.
}
\keyword{internal}
