% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_duplicate_cols.R
\name{drop_duplicate_cols}
\alias{drop_duplicate_cols}
\title{Drop duplicate columns}
\usage{
drop_duplicate_cols(dt)
}
\arguments{
\item{dt}{data.table}
}
\value{
Null output
}
\description{
Drop columns with identical names (if any exist) within a data.table.
}
\keyword{internal}
