% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_duplicate_rows.R
\name{drop_duplicate_rows}
\alias{drop_duplicate_rows}
\title{Drop duplicate rows}
\usage{
drop_duplicate_rows(dt, verbose = TRUE)
}
\arguments{
\item{dt}{data.table}

\item{verbose}{Print messages.}
}
\value{
Filtered \code{dt}.
}
\description{
Drop rows with duplicate values across all columns.
}
\keyword{internal}
