% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_to_granges.R
\name{dt_to_granges}
\alias{dt_to_granges}
\title{Convert data.table to GRanges object}
\usage{
dt_to_granges(
  dat,
  style = c("NCBI", "UCSC"),
  chrom_col = "CHR",
  start_col = "POS",
  end_col = "POS"
)
}
\arguments{
\item{dat}{data.table of genomic data}

\item{style}{conversion to GRanges style, NCBI or UCSC}

\item{chrom_col}{column for chromosome}

\item{start_col}{column for start position}

\item{end_col}{column for end position (is the same as start for snps)}
}
\value{
GRanges object of the genomic data
}
\description{
Convert data.table to GRanges object
}
\keyword{internal}
