% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sumstats.R
\name{format_sumstats}
\alias{format_sumstats}
\title{Check that summary statistics from GWAS are in a homogeneous format}
\usage{
format_sumstats(path)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
The address for the modified sumstats file
}
\description{
Check that summary statistics from GWAS are in a homogeneous format
}
\examples{
#format_sumstats(path)
}
