% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatted_example.R
\name{formatted_example}
\alias{formatted_example}
\title{Formatted example}
\usage{
formatted_example(
  path = system.file("extdata", "eduAttainOkbay.txt", package = "MungeSumstats"),
  formatted = TRUE,
  sorted = TRUE
)
}
\arguments{
\item{path}{Path to raw example file. Default to built-in dataset.}

\item{formatted}{Whether the column names should be formatted
(default:\code{TRUE}).}

\item{sorted}{Whether the rows should be sorted by genomic coordinates
(default:\code{TRUE}).}
}
\value{
\code{sumstats_dt}
}
\description{
Returns an example of summary stats that have had their column names
already standardised with
\link[MungeSumstats]{standardise_sumstats_column_headers_crossplatform}.
}
\examples{
sumstats_dt <- MungeSumstats::formatted_example()
}
