% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eff_frq_allele_combns.R
\name{get_eff_frq_allele_combns}
\alias{get_eff_frq_allele_combns}
\title{Get combinations of uncorrected allele and effect (and frq) columns}
\usage{
get_eff_frq_allele_combns(
  mapping_file = sumstatsColHeaders,
  eff_frq_cols = c("BETA", "OR", "LOG_ODDS", "SIGNED_SUMSTAT", "Z", "FRQ")
)
}
\arguments{
\item{mapping_file}{MungeSumstats has a pre-defined column-name mapping file
which should cover the most common column headers and their interpretations.
However, if a column header that is in youf file is missing of the mapping we
give is incorrect you can supply your own mapping file. Must be a 2 column
dataframe with column names "Uncorrected" and "Corrected". See
data(sumstatsColHeaders) for default mapping and necessary format.}

\item{eff_frq_cols}{Corrected effect or frequency column names found in a
sumstats. Default of BETA, OR, LOG_ODDS, SIGNED_SUMSTAT, Z and FRQ.}
}
\value{
datatable containing uncorrected and corrected combinations
}
\description{
Get combinations of uncorrected allele and effect (and frq) columns
}
