% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_build.R
\name{get_genome_build}
\alias{get_genome_build}
\title{Infers the genome build of the summary statistics file (GRCh37 or GRCh38)
from the data. Uses SNP (RS ID) & CHR & BP to get genome build.}
\usage{
get_genome_build(
  sumstats,
  nThread = 1,
  sampled_snps = 10000,
  standardise_headers = TRUE,
  mapping_file = sumstatsColHeaders
)
}
\arguments{
\item{sumstats}{data table/data frame obj of the summary statistics file for
the GWAS ,or file path to summary statistics file.}

\item{nThread}{Number of threads to use for parallel processes.}

\item{sampled_snps}{Downsample the number of SNPs used when inferring genome
build to save time.}

\item{standardise_headers}{Run
\code{standardise_sumstats_column_headers_crossplatform}.}

\item{mapping_file}{MungeSumstats has a pre-defined column-name mapping file
which should cover the most common column headers and their interpretations.
However, if a column header that is in youf file is missing of the mapping we
give is incorrect you can supply your own mapping file. Must be a 2 column
dataframe with column names "Uncorrected" and "Corrected". See
data(sumstatsColHeaders) for default mapping and necessary format.}
}
\value{
ref_genome the genome build of the data
}
\description{
Infers the genome build of the summary statistics file (GRCh37 or GRCh38)
from the data. Uses SNP (RS ID) & CHR & BP to get genome build.
}
\examples{
#Pass path to Educational Attainment Okbay sumstat file to a temp directory

eduAttainOkbayPth <- system.file("extdata","eduAttainOkbay.txt", 
                                 package="MungeSumstats")

## Call uses reference genome as default with more than 2GB of memory,
## which is more than what 32-bit Windows can handle so remove certain checks
is_32bit_windows <- 
.Platform$OS.type == "windows" && .Platform$r_arch == "i386"
if (!is_32bit_windows) {
ref_genome <- get_genome_build(sumstats=eduAttainOkbayPth)                                  
}
}
