% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_builds.R
\name{get_genome_builds}
\alias{get_genome_builds}
\title{Infer genome builds}
\usage{
get_genome_builds(
  sumstats_list,
  header_only = TRUE,
  sampled_snps = 10000,
  names_from_paths = FALSE,
  nThread = 1
)
}
\arguments{
\item{sumstats_list}{A named list of paths to summary statistics,
or a named list of \code{data.table} objects.}

\item{header_only}{Instead of reading in the entire \code{sumstats} file,
only read in the first N rows where N=\code{sampled_snps}.
This should help speed up cases where you have to read in \code{sumstats}
from disk each time.}

\item{sampled_snps}{Downsample the number of SNPs used when inferring genome
build to save time.}

\item{names_from_paths}{Infer the name of each item in \code{sumstats_list}
from its respective file path.
Only works if \code{sumstats_list} is a list of paths.}

\item{nThread}{Number of threads to use for parallel processes.}
}
\value{
ref_genome the genome build of the data
}
\description{
Infers the genome build of  summary statistics files (GRCh37 or GRCh38)
from the data. Uses SNP (RSID) & CHR & BP to get genome build.
}
\details{
Iterative version of \code{get_genome_build}.
}
\examples{
# Pass path to Educational Attainment Okbay sumstat file to a temp directory

eduAttainOkbayPth <- system.file("extdata", "eduAttainOkbay.txt",
    package = "MungeSumstats"
)
sumstats_list <- list(ss1 = eduAttainOkbayPth, ss2 = eduAttainOkbayPth)

## Call uses reference genome as default with more than 2GB of memory,
## which is more than what 32-bit Windows can handle so remove certain checks
is_32bit_windows <-
    .Platform$OS.type == "windows" && .Platform$r_arch == "i386"
if (!is_32bit_windows) {
    ref_genomes <- get_genome_builds(sumstats_list = sumstats_list)
}
}
