% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_name_log_file.R
\name{get_unique_name_log_file}
\alias{get_unique_name_log_file}
\title{Simple function to ensure the new entry name to a list doesn't have the same
name as another entry}
\usage{
get_unique_name_log_file(name, log_files)
}
\arguments{
\item{name}{proposed name for the entry}

\item{log_files}{list of log file locations}
}
\value{
a unique name (character)
}
\description{
Simple function to ensure the new entry name to a list doesn't have the same
name as another entry
}
\keyword{internal}
