% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vcf_sample_ids.R
\name{get_vcf_sample_ids}
\alias{get_vcf_sample_ids}
\title{Get VCF sample ID(s)}
\usage{
get_vcf_sample_ids(path)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted. Default
is a temporary file.}
}
\value{
Tsample_id
}
\description{
Get VCF sample ID(s)
}
\keyword{internal}
