% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_to_dt.R
\name{granges_to_dt}
\alias{granges_to_dt}
\title{GenomicRanges to data.table}
\source{
\href{https://rdrr.io/github/databio/GenomicDistributions/src/R/utility.R}{
Code adapted from GenomicDistributions.}
}
\usage{
granges_to_dt(gr)
}
\arguments{
\item{gr}{A \link[GenomicRanges]{GRanges} object.}
}
\value{
A data.table object.
}
\description{
Convert a \link[GenomicRanges]{GRanges} into a \link[data.table]{data.table}.
}
\keyword{internal}
