% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hg19ToHg38}
\alias{hg19ToHg38}
\title{UCSC Chain file hg19 to hg38}
\format{
gunzipped chain file
}
\source{
The chain file was downloaded from
https://hgdownload.cse.ucsc.edu/goldenpath/hg19/liftOver/
\code{
utils::download.file('ftp://hgdownload.cse.ucsc.edu/goldenPath/hg19/liftOver/hg19ToHg38.over.chain.gz',tempdir())
}
}
\description{
UCSC Chain file hg19 to hg38, .chain.gz file, downloaded from
https://hgdownload.cse.ucsc.edu/goldenpath/hg19/liftOver/ on 09/10/21
}
\details{
UCSC Chain file hg19 to hg38, .chain.gz file, downloaded on 09/10/21
To be used as a back up if the download from UCSC fails.
}
\section{hg19ToHg38.over.chain.gz}{
NA
}

