% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hg38ToHg19}
\alias{hg38ToHg19}
\title{UCSC Chain file hg38 to hg19}
\format{
gunzipped chain file
}
\source{
The chain file was downloaded from
https://hgdownload.cse.ucsc.edu/goldenpath/hg38/liftOver/
\code{
utils::download.file('ftp://hgdownload.cse.ucsc.edu/goldenPath/hg38/liftOver/hg38ToHg19.over.chain.gz',tempdir())
}
}
\description{
UCSC Chain file hg38 to hg19, .chain.gz file, downloaded from
https://hgdownload.cse.ucsc.edu/goldenpath/hg19/liftOver/ on 09/10/21
}
\details{
UCSC Chain file hg38 to hg19, .chain.gz file, downloaded on 09/10/21
To be used as a back up if the download from UCSC fails.
}
\section{hg38ToHg19.over.chain.gz}{
NA
}

