% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ieu-a-298}
\alias{ieu-a-298}
\title{Local ieu-a-298 file from IEU Open GWAS}
\format{
gunzipped tsv file
}
\source{
The file was downloaded with:
\code{
MungeSumstats::import_sumstats(ids = "ieu-a-298",ref_genome = "GRCH37")
}
}
\description{
Local ieu-a-298 file from IEU Open GWAS, downloaded on 09/10/21.
}
\details{
Local ieu-a-298 file from IEU Open GWAS, downlaoded on 09/10/21.
This is done in case the download in the package vignette fails.
}
\section{ieu-a-298.tsv.gz}{
NA
}

