% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ieuAmlVcf}
\alias{ieuAmlVcf}
\title{GWAS Amyotrophic lateral sclerosis ieu open GWAS project - Subset}
\format{
character vector with 528 items relating to 99 SNPs
}
\source{
The summary statistics VCF (VCFv4.2) file was downloaded from
https://gwas.mrcieu.ac.uk/datasets/ebi-a-GCST005647/
and formatted to a .rda with the following:
\code{
#Get example VCF dataset, use GWAS Amyotrophic lateral sclerosis
AML_GWAS_VCF <- readLines("ebi-a-GCST005647.vcf.gz")
#Subset to just the first 99 SNPs
ieuAmlVcf <- AML_GWAS_VCF[1:528]
usethis::use_data(ieuAmlVcf,overwrite = TRUE)
}
}
\usage{
ieuAmlVcf
}
\description{
VCF (VCFv4.2) of the GWAS Amyotrophic lateral sclerosis ieu open GWAS project
Dataset: ebi-a-GCST005647.
A subset of 99 SNPs
}
\keyword{datasets}
