% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sumstats.R
\name{import_sumstats}
\alias{import_sumstats}
\title{Import full genome-wide GWAS summary statistics from Open GWAS}
\usage{
import_sumstats(
  ids,
  vcf_download = FALSE,
  vcf_dir = tempdir(),
  force_new = FALSE,
  ...
)
}
\arguments{
\item{ids}{List of Open GWAS study IDs (e.g. \code{c("prot-a-664", "ieu-b-4760")}).}

\item{vcf_download}{Download the original VCF from Open GWAS.}

\item{vcf_dir}{Where to download the original VCF from Open GWAS.}

\item{force_new}{Overwrite a previously downloaded VCF with the same path name.}

\item{...}{Additional arguments to be passed to \code{MungeSumstats::format_sumstats}.}
}
\value{
Either a named list of data objects or paths, depending on the arguments passed to
}
\description{
Import full genome-wide GWAS summary statistics from Open GWAS
}
\examples{
### Search by criteria
metagwas <- MungeSumstats::find_sumstats()
### Only use a subset for testing purposes                                           
ids <- (dplyr::arrange(metagwas, nsnp))$id[1:2]                                     
datasets <- MungeSumstats::import_sumstats(ids = ids)                                           
}
