% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_chrom_format.R
\name{infer_chrom_format}
\alias{infer_chrom_format}
\title{Infer chrom format}
\usage{
infer_chrom_format(path, nrows, verbose = TRUE)
}
\arguments{
\item{path}{Path to VCF.}

\item{nrows}{Number of rows to sample.}

\item{verbose}{Print messages.}
}
\value{
logical: whether or not the chromosome contains "chr".
}
\description{
Infer chromosome format of a VCF file.
}
\keyword{internal}
