% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_vcf_sample_ids.R
\name{infer_vcf_sample_ids}
\alias{infer_vcf_sample_ids}
\title{Infer VCF sample ID(s)}
\usage{
infer_vcf_sample_ids(sample_id = NULL, sumstats_file)
}
\value{
sample_id
}
\description{
Infer VCF sample ID(s)
}
\keyword{internal}
