% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tabix.R
\name{is_tabix}
\alias{is_tabix}
\title{Is tabix}
\usage{
is_tabix(path)
}
\arguments{
\item{path}{Path to file.}
}
\value{
logical: whether the file is tabix-indexed or not.
}
\description{
Is a file bgz-zip compressed and tabix-indexed.
}
