% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieugwasr.R
\name{legacy_ids}
\alias{legacy_ids}
\title{Convert current IDs to legacy IDs}
\usage{
legacy_ids(x)
}
\arguments{
\item{x}{Vector of ids}
}
\value{
vector of back compatible ids
}
\description{
Convert current IDs to legacy IDs
}
\keyword{internal}
