% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{liftover}
\alias{liftover}
\title{Genome build liftover}
\source{
\href{https://doi.org/doi:10.18129/B9.bioc.liftOver}{liftOver}

\href{https://hgdownload.cse.ucsc.edu/goldenpath/hg19/liftOver/}{
UCSC chain files}
}
\usage{
liftover(
  sumstats_dt,
  convert_ref_genome,
  ref_genome,
  imputation_ind,
  verbose = TRUE
)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{convert_ref_genome}{name of the reference genome to convert to
("GRCh37" or "GRCh38"). This will only occur if the current genome build does
not match. Default is not to convert the genome build (NULL).}

\item{ref_genome}{name of the reference genome used for the GWAS ("GRCh37" or
"GRCh38"). Argument is case-insensitive. Default is NULL which infers the
reference genome from the data.}

\item{imputation_ind}{Binary Should a column be added for each imputation
step to show what SNPs have imputed values for differing fields. This
includes a field denoting SNP allele flipping (flipped). On the flipped
value, this denoted whether the alelles where switched based on
MungeSumstats initial choice of A1, A2 from the input column headers and thus
may not align with what the creator intended.\strong{Note} these columns will be
in the formatted summary statistics returned. Default is FALSE.}

\item{verbose}{Print messages.}
}
\value{
\code{list("sumstats_dt"=sumstats_dt)}
}
\description{
Transfer genomic coordinates from one genome build to another.
}
\keyword{internal}
