% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_snp_loc_data.R
\name{load_snp_loc_data}
\alias{load_snp_loc_data}
\title{Loads the SNP locations and alleles for Homo sapiens extracted from
NCBI dbSNP Build 144. Reference genome version is dependent on user input.}
\usage{
load_snp_loc_data(msg = NULL)
}
\arguments{
\item{msg}{Optional name of the column missing from the dataset in question}
}
\value{
SNP_LOC_DATA SNP positions and alleles for Homo sapiens extracted
from NCBI dbSNP Build 144
}
\description{
Loads the SNP locations and alleles for Homo sapiens extracted from
NCBI dbSNP Build 144. Reference genome version is dependent on user input.
}
\examples{
#SNP_LOC_DATA <- load_snp_loc_data()

}
