% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_allele_upper.R
\name{make_allele_upper}
\alias{make_allele_upper}
\title{Ensure A1 and A2 are upper case}
\usage{
make_allele_upper(sumstats_dt, log_files)
}
\arguments{
\item{log_files}{list of log file locations}
}
\value{
list containing sumstats_dt, the modified summary statistics data
table object and the log file list
}
\description{
Ensure A1 and A2 are upper case
}
\keyword{internal}
