% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messager.R
\name{messager}
\alias{messager}
\title{Print messages}
\usage{
messager(..., v = TRUE)
}
\arguments{
\item{...}{Message input.}

\item{v}{Whether to print messages.}
}
\value{
Null output.
}
\description{
Print messages with option to silence.
}
\keyword{internal}
