% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dropped_INFO.R
\name{parse_dropped_INFO}
\alias{parse_dropped_INFO}
\title{Parse number of SNPs dropped due to being below the INFO threshold}
\usage{
parse_dropped_INFO(l)
}
\arguments{
\item{l}{Lines of text from log file.}
}
\value{
Numeric
}
\description{
Support function for \link[MungeSumstats]{parse_logs}.
}
\keyword{internal}
