% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dropped_nonA1A2.R
\name{parse_dropped_nonA1A2}
\alias{parse_dropped_nonA1A2}
\title{Parse number of SNPs dropped due to not matching the ref genome A1 or A2}
\usage{
parse_dropped_nonA1A2(l)
}
\arguments{
\item{l}{Lines of text from log file.}
}
\value{
Numeric
}
\description{
Support function for \link[MungeSumstats]{parse_logs}.
}
\keyword{internal}
