% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dropped_nonRef.R
\name{parse_dropped_nonRef}
\alias{parse_dropped_nonRef}
\title{Parse number of SNPs dropped due to being in the ref genome}
\usage{
parse_dropped_nonRef(l)
}
\arguments{
\item{l}{Lines of text from log file.}
}
\value{
Numeric
}
\description{
Support function for \link[MungeSumstats]{parse_logs}.
}
\keyword{internal}
