% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_flipped.R
\name{parse_flipped}
\alias{parse_flipped}
\title{Parse number of SNPs flipped to align with the ref genome}
\usage{
parse_flipped(l)
}
\arguments{
\item{l}{Lines of text from log file.}
}
\value{
Numeric
}
\description{
Support function for \link[MungeSumstats]{parse_logs}.
}
\keyword{internal}
