% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_genome_build.R
\name{parse_genome_build}
\alias{parse_genome_build}
\title{Genome build inferred from the summary statistics}
\usage{
parse_genome_build(l)
}
\arguments{
\item{l}{Lines of text from log file.}
}
\value{
Character
}
\description{
Support function for \link[MungeSumstats]{parse_logs}.
}
\keyword{internal}
