% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_idStandard.R
\name{parse_idStandard}
\alias{parse_idStandard}
\title{Standardised IEU MRC OpenGWAS ID}
\usage{
parse_idStandard(l)
}
\arguments{
\item{l}{Lines of text from log file.}
}
\value{
Character
}
\description{
Support function for \link[MungeSumstats]{parse_logs}.
}
\keyword{internal}
