% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pval_neg.R
\name{parse_pval_neg}
\alias{parse_pval_neg}
\title{Parse number of SNPs with p-values <0}
\usage{
parse_pval_neg(l)
}
\arguments{
\item{l}{Lines of text from log file.}
}
\value{
Numeric
}
\description{
Support function for \link[MungeSumstats]{parse_logs}.
}
\keyword{internal}
