% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pval_small.R
\name{parse_pval_small}
\alias{parse_pval_small}
\title{Parse number of SNPs with non-negative p-values <=5e-324}
\usage{
parse_pval_small(l)
}
\arguments{
\item{l}{Lines of text from log file.}
}
\value{
Numeric
}
\description{
Support function for \link[MungeSumstats]{parse_logs}.
}
\keyword{internal}
