% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_report.R
\name{parse_report}
\alias{parse_report}
\title{Parse "Summary statistics report" metrics}
\usage{
parse_report(l, entry = 1, line = 1)
}
\arguments{
\item{l}{Lines of text from log file.}
}
\value{
Numeric
}
\description{
Support function for \link[MungeSumstats]{parse_logs}.
}
\keyword{internal}
