% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_snps_freq_05.R
\name{parse_snps_freq_05}
\alias{parse_snps_freq_05}
\title{Parse number/percent of SNPs with FREQ values >0.5}
\usage{
parse_snps_freq_05(l, percent = FALSE)
}
\arguments{
\item{l}{Lines of text from log file.}
}
\value{
Numeric
}
\description{
Support function for \link[MungeSumstats]{parse_logs}.
}
\keyword{internal}
