% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_snps_not_formatted.R
\name{parse_snps_not_formatted}
\alias{parse_snps_not_formatted}
\title{Parse number of SNPs not correctly formatted}
\usage{
parse_snps_not_formatted(l)
}
\arguments{
\item{l}{Lines of text from log file.}
}
\value{
Numeric
}
\description{
Support function for \link[MungeSumstats]{parse_logs}.
}
\keyword{internal}
