% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_time.R
\name{parse_time}
\alias{parse_time}
\title{Parse the total time taken the munge the file}
\usage{
parse_time(l)
}
\arguments{
\item{l}{Lines of text from log file.}
}
\value{
Character
}
\description{
Support function for \link[MungeSumstats]{parse_logs}.
}
\keyword{internal}
