% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_log_pval.R
\name{read_log_pval}
\alias{read_log_pval}
\title{Read -log10 p-value column}
\usage{
read_log_pval(
  sumstats_dt,
  mapping_file = sumstatsColHeaders,
  return_list = TRUE
)
}
\arguments{
\item{sumstats_dt}{Summary stats data.table.}

\item{mapping_file}{MungeSumstats has a pre-defined column-name mapping file
which should cover the most common column headers and their interpretations.
However, if a column header that is in youf file is missing of the mapping we
give is incorrect you can supply your own mapping file. Must be a 2 column
dataframe with column names "Uncorrected" and "Corrected". See
data(sumstatsColHeaders) for default mapping and necessary format.}

\item{return_list}{Binary, whether to return the dt in a list or not - list
is standard for the \code{format_sumstats()} function.}
}
\value{
Null output.
}
\description{
Parse p-value column in VCF file.of other general -loq10 p-values
}
\keyword{internal}
