% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sumstats.R
\name{read_sumstats}
\alias{read_sumstats}
\title{Determine sum stats file type and reda them into memory}
\usage{
read_sumstats(path, nThread = 1)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted}

\item{nThread}{Number of threads to use for parallel processes.}
}
\description{
Determine sum stats file type and reda them into memory
}
\keyword{internal}
