% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Read in VCF format}
\usage{
read_vcf(path, nThread = 1)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted}

\item{nThread}{Number of threads to use for parallel processes.}
}
\value{
The modified sumstats_file
}
\description{
Read in VCF format
}
\keyword{internal}
