% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf_genome.R
\name{read_vcf_genome}
\alias{read_vcf_genome}
\title{Read VCF genome}
\usage{
read_vcf_genome(
  header = NULL,
  validate = FALSE,
  default_genome = "HG19/GRCh37",
  verbose = TRUE
)
}
\arguments{
\item{header}{Header extracted by \link[VariantAnnotation]{scanVcfHeader}.}

\item{validate}{Validate genome name using
\link[GenomeInfoDb]{mapGenomeBuilds}.}

\item{default_genome}{When no genome can be extracted,
default to this genome build.}

\item{verbose}{Print messages.}
}
\value{
genome
}
\description{
Get the genome build of a remote or local VCF file.
}
\keyword{internal}
