% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf_info.R
\name{read_vcf_info}
\alias{read_vcf_info}
\title{Read VCF: INFO column}
\usage{
read_vcf_info(sumstats_dt)
}
\arguments{
\item{sumstats_dt}{Summary stats data.table.}
}
\value{
Null output.
}
\description{
Parse INFO column in VCF file.
}
\keyword{internal}
