% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf_markername.R
\name{read_vcf_markername}
\alias{read_vcf_markername}
\title{Read VCF: MarkerName column}
\usage{
read_vcf_markername(sumstats_dt)
}
\arguments{
\item{sumstats_dt}{Summary stats data.table.}
}
\value{
Null output.
}
\description{
Parse MarkerName/SNP column in VCF file.
}
\keyword{internal}
