% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf_pval.R
\name{read_vcf_pval}
\alias{read_vcf_pval}
\title{Read VCF: p-value column}
\usage{
read_vcf_pval(sumstats_dt)
}
\arguments{
\item{sumstats_dt}{Summary stats data.table.}
}
\value{
Null output.
}
\description{
Parse p-value column in VCF file.
}
\keyword{internal}
