% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_cores.R
\name{register_cores}
\alias{register_cores}
\title{Register cores}
\usage{
register_cores(workers = 1, progressbar = TRUE)
}
\arguments{
\item{workers}{

    \code{integer(1)} Number of workers. Defaults to the maximum of 1 or
    the number of cores determined by \code{detectCores} minus 2 unless
    environment variables \code{R_PARALLELLY_AVAILABLECORES_FALLBACK} or
    \code{BIOCPARALLEL_WORKER_NUMBER} are set otherwise. For a
    \code{SOCK} cluster, \code{workers} can be a \code{character()}
    vector of host names.

  }

\item{progressbar}{
    \code{logical(1)} Enable progress bar (based on plyr:::progress_text).
  }
}
\value{
Null output.
}
\description{
Register a multi-threaded instances using \pkg{BiocParallel}.
}
