% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_cols.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{Remove empty columns}
\usage{
remove_empty_cols(sumstats_dt, sampled_rows = NULL, verbose = TRUE)
}
\arguments{
\item{sampled_rows}{First N rows to sample.
Set \code{NULL} to use full \code{sumstats_file}.
when determining whether cols are empty.}

\item{verbose}{Print messages.}
}
\value{
Null output.
}
\description{
Remote columns that are empty or contain all the same values in a data.table.
}
\keyword{internal}
