% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_summary.R
\name{report_summary}
\alias{report_summary}
\title{Report info on current state of the summary statistics}
\usage{
report_summary(sumstats_dt, orig_dims = NULL)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary
statistics file for the GWAS.}
}
\value{
No return
}
\description{
Prints report.
}
\keyword{internal}
