% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieugwasr.R
\name{select_api}
\alias{select_api}
\title{Toggle API address between development and release}
\usage{
select_api(where = "public")
}
\arguments{
\item{where}{Which API to use. Choice between "local", "release", "test". Default = "local"}
}
\value{
No return
}
\description{
From \code{ieugwasr}.
}
\keyword{internal}
