% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_vcf_fields.R
\name{select_vcf_fields}
\alias{select_vcf_fields}
\title{Select VCF fields}
\usage{
select_vcf_fields(
  path,
  sampled_rows = 1e+07,
  which = NULL,
  single_sample = TRUE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted. A
dataframe or datatable of the summary statistics file can also be passed
directly to MungeSumstats using the path parameter.}

\item{sampled_rows}{First N rows to sample.
Set \code{NULL} to use full \code{sumstats_file}.
when determining whether cols are empty.}

\item{which}{A \code{\linkS4class{GRanges}} describing the sequences and
    ranges to be queried. Variants whose \code{POS} lies in the interval(s)
    \code{[start, end]} are returned. If \code{which} is not specified all
    ranges are returned. 
  }

\item{single_sample}{If multiple samples are present in the same VCF,
only return the first one (default: \code{TRUE}).}

\item{nThread}{Number of threads to use for parallel processes.}

\item{verbose}{Print messages.}
}
\value{
\code{ScanVcfParam} object.
}
\description{
Select non-empty columns from each VCF field type.
}
\keyword{internal}
