% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_coord_genomicranges.R
\name{sort_coord_genomicranges}
\alias{sort_coord_genomicranges}
\title{Sort sum stats: GenomicRanges}
\usage{
sort_coord_genomicranges(sumstats_dt)
}
\arguments{
\item{sumstats_dt}{\link[data.table]{data.table} obj of the
summary statistics file for the GWAS.}
}
\value{
Sorted sumstats_dt
}
\description{
Sort summary statistics table by genomic coordinates using a slower
(but in some cases more robust) \code{GenomicRanges} strategy
}
\keyword{internal}
