% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_coordinates.R
\name{sort_coords}
\alias{sort_coords}
\title{Sort sum stats by genomic coordinates}
\usage{
sort_coords(sumstats_dt, sort_coordinates = TRUE)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{sort_coords}{Whether to sort by coordinates.}
}
\description{
Sort sum stats by genomic coordinates
}
\keyword{internal}
