% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_coords_datatable.R
\name{sort_coords_datatable}
\alias{sort_coords_datatable}
\title{Sort sum stats: data.table}
\usage{
sort_coords_datatable(
  sumstats_dt,
  chr_col = "CHR",
  start_col = "BP",
  end_col = start_col
)
}
\arguments{
\item{sumstats_dt}{\link[data.table]{data.table} obj of the
summary statistics file for the GWAS.}

\item{chr_col}{Chromosome column name.}

\item{start_col}{Genomic end position column name.}
}
\value{
Sorted sumstats_dt
}
\description{
Sort summary statistics table by genomic coordinates using a fast
\code{data.table}-native strategy
}
\keyword{internal}
