% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumstats_reader.R
\name{sumstats_reader}
\alias{sumstats_reader}
\title{Determine sum stats file type and reda them into memory}
\usage{
sumstats_reader(path, nThread = 1)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted}

\item{nThread}{Number of threads to use for parallel processes.}
}
\description{
Determine sum stats file type and reda them into memory
}
\keyword{internal}
