% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_granges.R
\name{to_granges}
\alias{to_granges}
\title{Convert to \code{GRanges}}
\usage{
to_granges(sumstats_dt)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file
for the GWAS.}
}
\value{
\code{GRanges} object
}
\description{
Convert to \code{GRanges}
}
\keyword{internal}
