% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_vranges.R
\name{to_vranges}
\alias{to_vranges}
\title{Convert to \code{VRanges}}
\usage{
to_vranges(sumstats_dt)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics
file for the GWAS.}
}
\value{
\code{VRanges} object
}
\description{
Convert to \code{VRanges}
}
\keyword{internal}
