% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_dt.R
\name{unlist_dt}
\alias{unlist_dt}
\title{Unlist a data.table}
\usage{
unlist_dt(dt)
}
\arguments{
\item{dt}{data.table}
}
\value{
\code{dt} with list columns turned into vectors.
}
\description{
Identify columns that are lists and turn them into vectors.
}
\keyword{internal}
