% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2df.R
\name{vcf2df}
\alias{vcf2df}
\title{Function to convert VariantAnnotation CollapsedVCF/ExpandedVCF
objects to a data frame}
\source{
https://gist.github.com/zhujack/849b75f5a8305edaeca1001dfb9c3fe9
}
\usage{
vcf2df(v, expand = TRUE)
}
\description{
Function to convert VariantAnnotation CollapsedVCF/ExpandedVCF
objects to a data frame
}
\keyword{internal}
