% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sumstats.R
\name{write_sumstats}
\alias{write_sumstats}
\title{Write sum stats file to disk}
\usage{
write_sumstats(
  sumstats_dt,
  check_save_out,
  write_vcf = FALSE,
  tabix_index = FALSE,
  nThread = 1
)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS.}

\item{check_save_out}{Output of \code{check_save_path()}.}

\item{write_vcf}{Whether to write as VCF (TRUE) or tabular file (FALSE).}
}
\value{
\code{VRanges} object
}
\description{
Write sum stats file to disk
}
\keyword{internal}
