% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_allele_flip.R
\name{check_allele_flip}
\alias{check_allele_flip}
\title{Ensure A1 & A2 are correctly named, if GWAS constructed as Risk/nonrisk alleles this will need to be converted}
\usage{
check_allele_flip(sumstats_file, path, ref_genome, rsids, allele_flip_check)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}

\item{ref_genome}{name of the reference genome used for the GWAS (GRCh37 or GRCh38)}

\item{rsids}{datatable of snpsById, filtered to SNPs of interest if loaded already. Or else NULL}

\item{allele_flip_check}{Binary Should the allele columns be chacked against reference genome to infer if flipping is necessary. Default is TRUE}
}
\value{
The modified sumstats_file
}
\description{
Ensure A1 & A2 are correctly named, if GWAS constructed as Risk/nonrisk alleles this will need to be converted
}
