% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chr.R
\name{check_chr}
\alias{check_chr}
\title{Ensure all SNPs on specified chromosomes are removed}
\usage{
check_chr(sumstats_dt, path, rmv_chr, make_uppercase = TRUE)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}

\item{rmv_chr}{vector or character The chromosomes on which the SNPs should be removed.
Use NULL if no filtering necessary. Default is X, Y and mitochondrial.}

\item{make_uppercase}{Make X and Y chromosomes uppercase.}
}
\value{
list containing sumstats_dt, the modified summary statistics data table object
}
\description{
Ensure all SNPs on specified chromosomes are removed
}
\keyword{internal}
