% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chr.R
\name{check_chr}
\alias{check_chr}
\title{Ensure all SNPs on specified chromosomes are removed}
\usage{
check_chr(sumstats_file, path, rmv_chr)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}

\item{rmv_chr}{vector or character The chromosomes on which the SNPs should be removed. Use NULL if no filtering necessary. Default is X, Y and mitochondrial.}
}
\value{
The modified sumstats_file
}
\description{
Ensure all SNPs on specified chromosomes are removed
}
