% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dup_bp.R
\name{check_dup_bp}
\alias{check_dup_bp}
\title{Ensure all rows have unique positions, drop those that don't}
\usage{
check_dup_bp(sumstats_dt, path)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
list containing sumstats_dt, the modified summary statistics data table object
}
\description{
Ensure all rows have unique positions, drop those that don't
}
\keyword{internal}
