% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_four_step_col.R
\name{check_four_step_col}
\alias{check_four_step_col}
\title{Ensure that CHR:BP:A2:A1 aren't merged into 1 column}
\usage{
check_four_step_col(sumstats_dt, path)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
list containing sumstats_dt, the modified summary statistics data table object
}
\description{
Ensure that CHR:BP:A2:A1 aren't merged into 1 column
}
\keyword{internal}
