% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_info_score.R
\name{check_info_score}
\alias{check_info_score}
\title{Ensure all SNPs have info score above threshold}
\usage{
check_info_score(sumstats_dt, path, INFO_filter)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted}

\item{INFO_filter}{numeric The minimum value permissible of the imputation
information score (if present in sumstats file). Default 0.9.}
}
\value{
list containing sumstats_dt, the modified summary statistics data
table object
}
\description{
Ensure all SNPs have info score above threshold
}
\keyword{internal}
