% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ldsc_format.R
\name{check_ldsc_format}
\alias{check_ldsc_format}
\title{Ensures that parameters are compatible with LDSC format}
\source{
\href{https://github.com/bulik/ldsc}{LDSC GitHub}
}
\usage{
check_ldsc_format(ldsc_format, convert_n_int, allele_flip_check, compute_z)
}
\arguments{
\item{compute_z}{Whether to include/compute Z-score in \code{sumstats_dt}.}
}
\value{
Formatted summary statistics
}
\description{
Format summary statistics for direct input to
Linkage Disequilibrium SCore (LDSC) regression without the need
to use their \code{munge_sumstats.py} script first.
}
\details{
\href{https://github.com/bulik/ldsc/wiki/Summary-Statistics-File-Format}{LDSC documentation}.
}
