% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_miss_data.R
\name{check_miss_data}
\alias{check_miss_data}
\title{Ensure that only one model in GWAS sumstats or only one trait tested}
\usage{
check_miss_data(sumstats_dt, path)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
list containing sumstats_dt, the modified summary statistics data table object
}
\description{
Ensure that only one model in GWAS sumstats or only one trait tested
}
\keyword{internal}
