% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_miss_data.R
\name{check_miss_data}
\alias{check_miss_data}
\title{Ensure that only one model in GWAS sumstats or only one trait tested}
\usage{
check_miss_data(
  sumstats_dt,
  path,
  log_folder_ind,
  check_save_out,
  tabix_index,
  nThread,
  log_files
)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted. Default
is a temporary file.}

\item{log_folder_ind}{Binary Should log files be stored containing all
filtered out SNPs (separate file per filter). The data is outputted in the
same format specified for the resulting sumstats file. The only exception to
this rule is if output is vcf, then log file saved as .tsv.gz. Default is
FALSE.}

\item{tabix_index}{Index the formatted summary statistics with
\href{http://www.htslib.org/doc/tabix.html}{tabix} for fast querying.}

\item{nThread}{Number of threads to use for parallel processes.}

\item{log_files}{list of log file locations}
}
\value{
list containing sumstats_dt, the modified summary statistics data
table object and a log file list.
}
\description{
Ensure that only one model in GWAS sumstats or only one trait tested
}
\keyword{internal}
