% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_multi_gwas.R
\name{check_multi_gwas}
\alias{check_multi_gwas}
\title{Ensure that only one model in GWAS sumstats or only one trait tested}
\usage{
check_multi_gwas(sumstats_file, path)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
The modified sumstats_file
}
\description{
Ensure that only one model in GWAS sumstats or only one trait tested
}
