% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_int.R
\name{check_n_int}
\alias{check_n_int}
\title{Ensure that the p values are not 3e-400 or lower, if so set to 0}
\usage{
check_n_int(sumstats_file, path)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
The modified sumstats_file
}
\description{
Ensure that the p values are not 3e-400 or lower, if so set to 0
}
