% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_int.R
\name{check_n_int}
\alias{check_n_int}
\title{Ensure that the N column are all integers}
\usage{
check_n_int(sumstats_file, path)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
The modified sumstats_file
}
\description{
Ensure that the N column are all integers
}
