% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_no_allele.R
\name{check_no_allele}
\alias{check_no_allele}
\title{Ensure that A1 & A2 are present, if not can find it with SNP and other allele}
\usage{
check_no_allele(sumstats_dt, path, ref_genome, rsids)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}

\item{ref_genome}{name of the reference genome used for the GWAS (GRCh37 or GRCh38)}

\item{rsids}{datatable of snpsById, filtered to SNPs of interest if loaded already. Or else NULL}
}
\value{
A list containing two data tables:
\itemize{
\item \code{sumstats_dt}: the modified summary statistics data table object
\item \code{rsids}: snpsById, filtered to SNPs of interest if loaded already. Or else NULL
}
}
\description{
More care needs to be taken if one of A1/A2 is present, before imputing the
other allele flipping needs to be checked
}
\keyword{internal}
