% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_on_ref_genome.R
\name{check_on_ref_genome}
\alias{check_on_ref_genome}
\title{Ensure all SNPs are on the reference genome}
\usage{
check_on_ref_genome(sumstats_dt, path, ref_genome, on_ref_genome, rsids)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}

\item{ref_genome}{name of the reference genome used for the GWAS (GRCh37 or GRCh38). Default is GRCh37.}

\item{on_ref_genome}{Binary Should a check take place that all SNPs are on the reference genome by SNP ID. Default is TRUE}

\item{rsids}{datatable of snpsById, filtered to SNPs of interest if loaded already. Or else NULL}
}
\value{
A list containing two data tables:
\itemize{
\item \code{sumstats_dt}: the modified summary statistics data table object
\item \code{rsids}: snpsById, filtered to SNPs of interest if loaded already. Or else NULL
}
}
\description{
Ensure all SNPs are on the reference genome
}
\keyword{internal}
