% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_save_path.R
\name{check_save_path}
\alias{check_save_path}
\title{Check if save path is appropriate}
\usage{
check_save_path(save_path)
}
\arguments{
\item{save_path}{File path to save formatted data. Defaults to \code{paste0(tempfile(),".tsv.gz")}.}
}
\description{
Returns corrected \code{save_path}, the file type, and the separator.
}
\keyword{internal}
