% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_signed_col.R
\name{check_signed_col}
\alias{check_signed_col}
\title{Ensure that there is at least one signed column in summary statistics file}
\usage{
check_signed_col(sumstats_file)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}
}
\value{
null
}
\description{
Ensure that there is at least one signed column in summary statistics file
}
