% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_small_p_val.R
\name{check_small_p_val}
\alias{check_small_p_val}
\title{Ensure that the p values are not 5e-324 or lower, if so set to 0}
\usage{
check_small_p_val(sumstats_dt, path, convert_small_p)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}

\item{convert_small_p}{Binary, should p-values < 5e-324 be converted to 0? Small p-values pass the R limit and can cause errors with LDSC/MAGMA and should be converted. Default is TRUE.}
}
\value{
list containing sumstats_dt, the modified summary statistics data table object
}
\description{
Ensure that the p values are not 5e-324 or lower, if so set to 0
}
\keyword{internal}
