% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_strand_ambiguous.R
\name{check_strand_ambiguous}
\alias{check_strand_ambiguous}
\title{Remove SNPs with strand-ambiguous alleles}
\usage{
check_strand_ambiguous(sumstats_dt, path, ref_genome, strand_ambig_filter)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}

\item{ref_genome}{name of the reference genome used for the GWAS (GRCh37 or GRCh38). Default is GRCh37.}

\item{strand_ambig_filter}{Binary Should SNPs with strand-ambiguous alleles be removed. Default is FALSE}
}
\value{
list containing sumstats_dt, the modified summary statistics data table object
}
\description{
Remove SNPs with strand-ambiguous alleles
}
\keyword{internal}
