% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tab_delimited.R
\name{check_tab_delimited}
\alias{check_tab_delimited}
\title{Ensure that tabs separate rows}
\usage{
check_tab_delimited(sumstats_file)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}
}
\value{
The modified sumstats_file
}
\description{
Ensure that tabs separate rows
}
\keyword{internal}
