% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vcf.R
\name{check_vcf}
\alias{check_vcf}
\title{Check if the inputted file is in VCF format, if so reformat as standard}
\usage{
check_vcf(sumstats_file, path)
}
\arguments{
\item{sumstats_file}{The summary statistics file for the GWAS}

\item{path}{Filepath for the summary statistics file to be formatted}
}
\value{
The modified sumstats_file
}
\description{
Check if the inputted file is in VCF format, if so reformat as standard
}
\keyword{internal}
