% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nsize.R
\name{compute_nsize}
\alias{compute_nsize}
\title{Check for N column if not present and user wants, impute N based on user's
sample size. \strong{NOTE} this will be the same value for each SNP which is not
necessarily correct and may cause issues down the line.}
\usage{
compute_nsize(sumstats_dt, imputation_ind, compute_n = TRUE)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the
GWAS.}

\item{imputation_ind}{Binary Should a column be added for each imputation
step to show what SNPs have imputed values for differing fields. This
includes a field denoting SNP allele flipping (flipped). \strong{Note}
these columns will be in the formatted summary statistics returned. Default
is FALSE.}

\item{compute_n}{Whether to impute N. Default of 0 won't impute, any other
integer will be imputed as the N (sample size) for every SNP in the dataset.
\strong{Note} that imputing the sample size for every SNP is not correct and}
}
\value{
\code{list("sumstats_dt"=sumstats_dt)}
}
\description{
Check for N column if not present and user wants, impute N based on user's
sample size. \strong{NOTE} this will be the same value for each SNP which is not
necessarily correct and may cause issues down the line.
}
\keyword{internal}
