% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sumstats.R
\name{convert_sumstats}
\alias{convert_sumstats}
\title{Convert summary statistics to desired object type}
\usage{
convert_sumstats(
  sumstats_dt,
  return_format = c("data.table", "vranges", "granges")
)
}
\arguments{
\item{return_format}{Object type to convert to;
\code{"data.table"}, \code{"GenomicRanges"} or
\code{"VRanges"}(default is \code{"data.table"}).}
}
\value{
Summary statistics in the converted format
}
\description{
Convert summary statistics to desired object type
}
\keyword{internal}
