% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sumstats.R
\name{format_sumstats}
\alias{format_sumstats}
\title{Check that summary statistics from GWAS are in a homogeneous format}
\usage{
format_sumstats(
  path,
  ref_genome = "GRCh37",
  convert_small_p = TRUE,
  convert_n_int = TRUE,
  analysis_trait = NULL
)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted}

\item{ref_genome}{name of the reference genome used for the GWAS (GRCh37 or GRCh38). Default is GRCh37.}

\item{convert_small_p}{Binary, should p-values < 5e-324 be converted to 0? Small p-values pass the R limit and can cause errors with LDSC/MAGMA and should be converted. Default is TRUE.}

\item{convert_n_int}{Binary, if N (the number of samples) is not an integer, should this be rounded? Default is TRUE.}

\item{analysis_trait}{If multiple traits were studied, name of the trait for analysis from the GWAS. Default is NULL}
}
\value{
The address for the modified sumstats file
}
\description{
Check that summary statistics from GWAS are in a homogeneous format
}
\examples{
#save the Educational Attainment Okbay sumstat file to a temp directory
tmp <- tempfile()
writeLines(MungeSumstats::eduAttainOkbay,con = tmp)
#pass path to format_sumstats
reformatted <- MungeSumstats::format_sumstats(tmp,ref_genome="GRCh37")
#returned location has the updated summary statistics file
}
