% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sumstats.R
\name{format_sumstats}
\alias{format_sumstats}
\title{Check that summary statistics from GWAS are in a homogeneous format}
\usage{
format_sumstats(
  path,
  ref_genome = NULL,
  convert_small_p = TRUE,
  compute_z = FALSE,
  convert_n_int = TRUE,
  analysis_trait = NULL,
  INFO_filter = 0.9,
  N_std = 5,
  N_dropNA = TRUE,
  rmv_chr = c("X", "Y", "MT"),
  rmv_chrPrefix = TRUE,
  on_ref_genome = TRUE,
  strand_ambig_filter = FALSE,
  allele_flip_check = TRUE,
  bi_allelic_filter = TRUE,
  sort_coordinates = TRUE,
  nThread = 1,
  save_path = tempfile(fileext = ".tsv.gz"),
  write_vcf = FALSE,
  tabix_index = FALSE,
  return_data = FALSE,
  return_format = "data.table",
  ldsc_format = FALSE,
  force_new = FALSE
)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted}

\item{ref_genome}{name of the reference genome used for the GWAS ("GRCh37" or "GRCh38").
Argument is case-insensitive. Default is NULL which infers the reference genome from the data.}

\item{convert_small_p}{Binary, should p-values < 5e-324 be converted to 0? Small p-values pass the R limit and can cause errors with LDSC/MAGMA and should be converted. Default is TRUE.}

\item{compute_z}{Whether to compute Z-score column from P.}

\item{convert_n_int}{Binary, if N (the number of samples) is not an integer, should this be rounded? Default is TRUE.}

\item{analysis_trait}{If multiple traits were studied, name of the trait for analysis from the GWAS. Default is NULL.}

\item{INFO_filter}{numeric The minimum value permissible of the imputation information score (if present in sumstatsfile). Default 0.9.}

\item{N_std}{numeric The number of standard deviations above the mean a SNP's N is needed to be removed. Default is 5.}

\item{N_dropNA}{Drop rows where N is missing.}

\item{rmv_chr}{vector or character The chromosomes on which the SNPs should be removed. Use NULL if no filtering necessary.
Default is X, Y and mitochondrial.}

\item{rmv_chrPrefix}{Remove "chr" or "CHR" from chromosome names.}

\item{on_ref_genome}{Binary Should a check take place that all SNPs are on the reference genome by SNP ID. Default is TRUE.}

\item{strand_ambig_filter}{Binary Should SNPs with strand-ambiguous alleles be removed. Default is FALSE.}

\item{allele_flip_check}{Binary Should the allele columns be checked against reference genome to infer if flipping is necessary. Default is TRUE.}

\item{bi_allelic_filter}{Binary Should non-biallelic SNPs be removed. Default is TRUE.}

\item{sort_coordinates}{Whether to sort by coordinates.}

\item{nThread}{Number of threads to use for parallel processes.}

\item{save_path}{File path to save formatted data. Defaults to \code{tempfile(fileext=".tsv.gz")}.}

\item{write_vcf}{Whether to write as VCF (TRUE) or tabular file (FALSE).}

\item{tabix_index}{Index the formatted summary statistics with \href{http://www.htslib.org/doc/tabix.html}{tabix} for fast querying.}

\item{return_data}{Return \code{data.table} directly to user. Otherwise, return the path to the save data. Default is FALSE.}

\item{return_format}{Object type to convert to;
\code{"data.table"}, \code{"GenomicRanges"} or
\code{"VRanges"}(default is \code{"data.table"}).}

\item{ldsc_format}{Ensure that output format meets all requirements
to be fed directly into LDSC without the need for additional munging.}

\item{force_new}{If a formatted file of the same names as \code{save_path} exists,
formatting will be skipped and this file will be imported instead (default).
Set \code{force_new=TRUE} to override this.}
}
\value{
The address for the modified sumstats file
}
\description{
Check that summary statistics from GWAS are in a homogeneous format
}
\examples{
#Pass path to Educational Attainment Okbay sumstat file to a temp directory

eduAttainOkbayPth <- system.file("extdata","eduAttainOkbay.txt", 
                                 package="MungeSumstats")

## Call uses reference genome as default with more than 2GB of memory,
## which is more than what 32-bit Windows can handle so remove certain checks

is_32bit_windows <- .Platform$OS.type == "windows" && .Platform$r_arch == "i386"
if (!is_32bit_windows) {
reformatted <- MungeSumstats::format_sumstats(path=eduAttainOkbayPth,
                                              ref_genome="GRCh37")
} else{
reformatted <- MungeSumstats::format_sumstats(path=eduAttainOkbayPth,
                                              ref_genome="GRCh37",
                                              on_ref_genome = FALSE,
                                              strand_ambig_filter=FALSE,
                                              bi_allelic_filter=FALSE,
                                              allele_flip_check=FALSE)
}
#returned location has the updated summary statistics file
}
