% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chain_file.R
\name{get_chain_file}
\alias{get_chain_file}
\title{Download chain file for liftover}
\source{
\href{https://hgdownload.cse.ucsc.edu/goldenpath/hg19/liftOver/}{
UCSC chain files}
}
\usage{
get_chain_file(
  build_conversion = c("hg38ToHg19", "hg19ToHg38"),
  save_dir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{build_conversion}{converting from what build to what? hg38ToHg19 or
hg19ToHg38}

\item{save_dir}{where is the chain file saved? Default is a temp directory}

\item{verbose}{extra messages printed? Default is TRUE}
}
\value{
loaded chain file for liftover
}
\description{
Download chain file for liftover
}
\keyword{internal}
