% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sumstats.R
\name{import_sumstats}
\alias{import_sumstats}
\title{Import full genome-wide GWAS summary statistics from Open GWAS}
\usage{
import_sumstats(
  ids,
  vcf_dir = tempdir(),
  vcf_download = TRUE,
  download_method = "download.file",
  quiet = TRUE,
  force_new = FALSE,
  nThread = 1,
  ...
)
}
\arguments{
\item{ids}{List of Open GWAS study IDs (e.g. \code{c("prot-a-664", "ieu-b-4760")}).}

\item{vcf_dir}{Where to download the original VCF from Open GWAS.}

\item{vcf_download}{Download the original VCF from Open GWAS.}

\item{download_method}{\code{"axel"} (multi-threaded) or \code{"download.file"} (single-threaded) .}

\item{quiet}{Run quietly.}

\item{force_new}{Overwrite a previously downloaded VCF with the same path name.}

\item{nThread}{Number of threads to use for parallel processes.}

\item{...}{Additional arguments to be passed to \code{MungeSumstats::format_sumstats}.}
}
\value{
Either a named list of data objects or paths, depending on the arguments passed to
}
\description{
Import full genome-wide GWAS summary statistics from Open GWAS
}
\examples{
### Search by criteria
metagwas <- MungeSumstats::find_sumstats()
### Only use a subset for testing purposes                                           
ids <- (dplyr::arrange(metagwas, nsnp))$id[1:2]    

### Default usage   
# datasets <- MungeSumstats::import_sumstats(ids = ids)
                                
#### Speed up with multi-threaded download via axel 
# datasets <- MungeSumstats::import_sumstats(ids = ids, vcf_download=TRUE, download_method="axel", nThread=10)                                           

}
