% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sumstats.R
\name{import_sumstats}
\alias{import_sumstats}
\title{Import full genome-wide GWAS summary statistics from Open GWAS}
\usage{
import_sumstats(
  ids,
  vcf_dir = tempdir(),
  vcf_download = TRUE,
  save_dir = tempdir(),
  write_vcf = FALSE,
  download_method = "download.file",
  quiet = TRUE,
  force_new = FALSE,
  nThread = 1,
  parallel_across_ids = FALSE,
  ...
)
}
\arguments{
\item{ids}{List of Open GWAS study IDs (e.g. \code{c("prot-a-664", "ieu-b-4760")}).}

\item{vcf_dir}{Where to download the original VCF from Open GWAS.
\emph{WARNING:} This is set to \code{tempdir()} by default.
This means the raw (pre-formatted) VCFs be deleted upon ending the R session.
Change this to keep the raw VCF file on disk (e.g. \code{vcf_dir="./raw_vcf"}).}

\item{vcf_download}{Download the original VCF from Open GWAS.}

\item{save_dir}{Directory to save formatted summary statistics in.}

\item{write_vcf}{Whether to write as VCF (TRUE) or tabular file (FALSE).}

\item{download_method}{\code{"axel"} (multi-threaded) or \code{"download.file"} (single-threaded) .}

\item{quiet}{Run quietly.}

\item{force_new}{Overwrite a previously downloaded VCF with the same path name.}

\item{nThread}{Number of threads to use for parallel processes.}

\item{parallel_across_ids}{If \code{parallel_across_ids=TRUE} and \code{nThread>1},
then each ID in \code{ids} will be processed in parallel.}

\item{...}{Additional arguments to be passed to \code{MungeSumstats::format_sumstats}.}
}
\value{
Either a named list of data objects or paths,
depending on the arguments passed to \code{format_sumstats}.
}
\description{
Requires internet access to run.
}
\examples{
### Search by criteria
metagwas <- MungeSumstats::find_sumstats(traits = c("parkinson","alzheimer"),
                                         min_sample_size = 5000)
### Only use a subset for testing purposes                                           
ids <- (dplyr::arrange(metagwas, nsnp))$id    

### Default usage   
## You can supply \code{import_sumstats()}
## with a list of as many OpenGWAS IDs as you want, 
## but we'll just give one to save time.

## Call uses reference genome as default with more than 2GB of memory,
## which is more than what 32-bit Windows can handle so remove certain checks

is_32bit_windows <- .Platform$OS.type == "windows" && .Platform$r_arch == "i386"
if (!is_32bit_windows) {
datasets <- MungeSumstats::import_sumstats(ids = ids[1]) 
} else{
reformatted <- MungeSumstats::import_sumstats(ids = ids[1],
                                              ref_genome="GRCh37",
                                              on_ref_genome = FALSE,
                                              strand_ambig_filter=FALSE,
                                              bi_allelic_filter=FALSE,
                                              allele_flip_check=FALSE)
} 

}
