% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_tabular.R
\name{index_tabular}
\alias{index_tabular}
\title{Tabix-index file: table}
\source{
Borrowed function from
\href{https://github.com/RajLabMSSM/echotabix/blob/main/R/convert.R}{
echotabix}.
}
\usage{
index_tabular(
  path,
  chrom_col = "CHR",
  start_col = "BP",
  end_col = start_col,
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to GWAS summary statistics file.}

\item{chrom_col}{column for chromosome}

\item{start_col}{column for start position}

\item{end_col}{column for end position (is the same as start for snps)}

\item{verbose}{Print messages.}
}
\value{
Path to tabix-indexed tabular file
}
\description{
Convert summary stats file to tabix format.
}
\examples{
eduAttainOkbayPth <- system.file("extdata", "eduAttainOkbay.txt",
                                 package = "MungeSumstats")
sumstats_dt <- data.table::fread(eduAttainOkbayPth, nThread = 1)
sumstats_dt <- 
MungeSumstats:::standardise_sumstats_column_headers_crossplatform(
    sumstats_dt = sumstats_dt)$sumstats_dt
sumstats_dt <- MungeSumstats:::sort_coords(sumstats_dt = sumstats_dt)
path <- tempfile(fileext = ".tsv")
MungeSumstats::write_sumstats(sumstats_dt = sumstats_dt, save_path = path)
    
indexed_file <- MungeSumstats::index_tabular(path = path)
}
\seealso{
Other tabix: 
\code{\link{index_vcf}()}
}
\concept{tabix}
