% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_snp_loc_data.R
\name{load_snp_loc_data}
\alias{load_snp_loc_data}
\title{Downloads and loads the SNP_LOC_DATA file that I stashed at figshare}
\usage{
load_snp_loc_data()
}
\value{
SNP_LOC_DATA Has four columns with SNP CHR BP and Build
}
\description{
Downloads and loads the SNP_LOC_DATA file that I stashed at figshare
}
\examples{
SNP_LOC_DATA <- load_snp_loc_data()

}
