% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_snp_loc_data.R
\name{load_snp_loc_data}
\alias{load_snp_loc_data}
\title{Loads the SNP locations and alleles for Homo sapiens from dbSNP builds}
\usage{
load_snp_loc_data(ref_genome, dbSNP, dbSNP_tarball = NULL, msg = NULL)
}
\arguments{
\item{ref_genome}{character, "GRCh37" or "GRCh38"}

\item{dbSNP}{integer, dbSNP build number (144, 155, or any installed
SNPlocs package)}

\item{dbSNP_tarball}{Optional path to a .tar.gz containing:
one or more .rds files (Bioc SNPlocs package layout).}

\item{msg}{optional character to message before loading}
}
\value{
A data.table or OnDiskLongTable of SNP locations
}
\description{
Loads the SNP locations and alleles for Homo sapiens from dbSNP builds
}
