% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_sumstats.R
\name{preview_sumstats}
\alias{preview_sumstats}
\title{Preview formatted sum stats saved to disk}
\usage{
preview_sumstats(save_path, nrows = 5L)
}
\arguments{
\item{save_path}{File path to save formatted data. Defaults to \code{paste0(tempfile(),".tsv.gz")}.}
}
\description{
Preview formatted sum stats saved to disk
}
\keyword{internal}
