% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_header.R
\name{read_header}
\alias{read_header}
\title{Read in file header}
\usage{
read_header(path, n = 2, skip_vcf_metadata = FALSE)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted}
}
\description{
Read in file header
}
\keyword{internal}
