% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_header.R
\name{read_header}
\alias{read_header}
\title{Read in file header}
\usage{
read_header(path, n = 2, skip_vcf_metadata = FALSE)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted. Default
is a temporary file.}

\item{n}{integer. The (maximal) number of lines to read. Negative values
indicate that one should read up to the end of input on the connection.}

\item{skip_vcf_metadata}{logical, should VCF metadata be ignored}
}
\value{
First \code{n} lines of the VCF header
}
\description{
Read in file header
}
\keyword{internal}
