% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sumstats.R
\name{read_sumstats}
\alias{read_sumstats}
\title{Determine summary statistics file type and reda them into memory}
\usage{
read_sumstats(path, nThread = 1)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted. Default
is a temporary file.}

\item{nThread}{Number of threads to use for parallel processes.}
}
\value{
\code{data.table} of formatted summary statistics
}
\description{
Determine summary statistics file type and reda them into memory
}
\examples{
 eduAttainOkbay <- MungeSumstats::read_sumstats(path = system.file("extdata","eduAttainOkbay.txt",
                                                                   package="MungeSumstats"))
}
