% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_cores.R
\name{register_cores}
\alias{register_cores}
\title{Register cores}
\usage{
register_cores(workers = 1, progressbar = TRUE)
}
\arguments{
\item{workers}{

    \code{integer(1)} Number of workers. Defaults to all cores available as
    determined by \code{detectCores}. For a \code{SOCK} cluster \code{workers}
    can be a \code{character()} vector of host names.

  }

\item{progressbar}{
    \code{logical(1)} Enable progress bar (based on plyr:::progress_text).
  }
}
\value{
Null output.
}
\description{
Register a multi-threaded instances using \pkg{BiocParallel}.
}
