% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_nonstandard_vcf_cols.R
\name{remove_nonstandard_vcf_cols}
\alias{remove_nonstandard_vcf_cols}
\title{Remove non-standard VCF columns}
\usage{
remove_nonstandard_vcf_cols(sample_id, sumstats_file, keep_extra_cols = FALSE)
}
\arguments{
\item{keep_extra_cols}{logical, should extra, non-essential columns from
input be kept}
}
\value{
sumstats_file
}
\description{
Remove non-standard VCF columns
}
\keyword{internal}
