% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumstatsColHeaders}
\alias{sumstatsColHeaders}
\title{Summary Statistics Column Headers}
\format{
dataframe with 82 rows nd 2 columns
}
\source{
The code to prepare the .Rda file file from the marker file is:
\code{
# Most the data in the below table comes from the LDSC github wiki
sumstatsColHeaders <- read.csv("inst/extdata/Magma_Column_headers.csv",
     stringsAsFactors = FALSE)
usethis::use_data(sumstatsColHeaders,overwrite = TRUE, internal=TRUE)
}
}
\usage{
data("sumstatsColHeaders")
}
\description{
List of uncorrected column headers often found in GWAS Summary
Statistics column headers
}
\keyword{datasets}
