% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_VRanges.R
\name{to_VRanges}
\alias{to_VRanges}
\title{Convert to \code{VRanges}}
\usage{
to_VRanges(sumstats_dt)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS.}
}
\value{
\code{VRanges} object
}
\description{
Convert to \code{VRanges}
}
\keyword{internal}
