% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_parameters.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Ensure that the input parameters are logical}
\usage{
validate_parameters(
  path,
  ref_genome,
  convert_small_p,
  convert_n_int,
  analysis_trait
)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted}

\item{ref_genome}{name of the reference genome used for the GWAS (GRCh37 or GRCh38)}

\item{convert_small_p}{Binary, should p-values < 5e-324 be converted to 0? Small p-values pass the R limit and can cause errors with LDSC/MAGMA and should be converted. Default is TRUE.}

\item{convert_n_int}{Binary, if N (the number of samples) is not an integer, should this be rounded? Default is TRUE.}

\item{analysis_trait}{If multiple traits were studied, name of the trait for analysis from the GWAS. Default is NULL}
}
\value{
No return
}
\description{
Ensure that the input parameters are logical
}
