% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2df.R
\name{vcf2df}
\alias{vcf2df}
\title{VCF to DF}
\source{
https://gist.github.com/zhujack/849b75f5a8305edaeca1001dfb9c3fe9

\code{
#### VariantAnnotation ####
# path <- "https://github.com/brentp/vcfanno/raw/master/example/exac.vcf.gz"
path <- system.file("extdata", "ALSvcf.vcf",
                    package = "MungeSumstats")
vcf <- VariantAnnotation::readVcf(file = path)
vcf_df <- MungeSumstats:::vcf2df(vcf = vcf)
}

{
\subsection{vcfR}{

if(!require("pinfsc50")) install.packages("pinfsc50")
vcf_file <- system.file("extdata", "pinf_sc50.vcf.gz", package = "pinfsc50")
vcf <- read.vcfR( vcf_file, verbose = FALSE )
vcf_df_list <- vcfR::vcfR2tidy(vcf, single_frame=TRUE)
vcf_df <- data.table::data.table(vcf_df_list$dat)
}
}
}
\usage{
vcf2df(vcf, add_sample_names = TRUE, add_rowranges = TRUE, verbose = TRUE)
}
\arguments{
\item{vcf}{Variant Call Format (VCF) file imported into R
as a \pkg{VariantAnnotation}
\link[VariantAnnotation]{CollapsedVCF}/
\link[VariantAnnotation]{ExpandedVCF} object.}

\item{add_sample_names}{Append sample names to column names
(e.g. "EZ" --> "EZ_ubm-a-2929").}

\item{add_rowranges}{Include \code{rowRanges} from VCF as well.}

\item{verbose}{Print messages.}
}
\value{
data.frame version of VCF
}
\description{
Function to convert a \pkg{VariantAnnotation}
\code{CollapsedVCF}/\code{ExpandedVCF}
object to a \code{data.frame}.
}
\keyword{internal}
