% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2df.R
\name{vcf2df}
\alias{vcf2df}
\title{VCF to DF}
\source{
https://gist.github.com/zhujack/849b75f5a8305edaeca1001dfb9c3fe9

\code{
vcf_file <- system.file("extdata", "ALSvcf.vcf",
                        package = "MungeSumstats")
vcf <- VariantAnnotation::readVcf(file = vcf_file)
vcf_df <- MungeSumstats::vcf2df(vcf = vcf)
}
}
\usage{
vcf2df(vcf, add_sample_names = TRUE)
}
\arguments{
\item{vcf}{Variant Call Format (VCF) file imported into R
as a \pkg{VariantAnnotation}
\link[VariantAnnotation]{CollapsedVCF}/
\link[VariantAnnotation]{ExpandedVCF} object.}

\item{add_sample_names}{Append sample names to column names
(e.g. "EZ" --> "EZ_ubm-a-2929").}
}
\value{
data.frame version of VCF
}
\description{
Function to convert a \pkg{VariantAnnotation}
\code{CollapsedVCF}/\code{ExpandedVCF}
object to a \code{data.frame}.
}
\keyword{internal}
