% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2df.R
\name{vcf2df}
\alias{vcf2df}
\title{VCF to DF}
\source{
https://gist.github.com/zhujack/849b75f5a8305edaeca1001dfb9c3fe9

\code{
vcf_file <- system.file("extdata", "ALSvcf.vcf",
                        package = "MungeSumstats")
vcf <- VariantAnnotation::readVcf(file = vcf_file)
vcf_df <- vcf2df(vcf = vcf)
}
}
\usage{
vcf2df(vcf, expand = TRUE)
}
\value{
data.frame version of VCF
}
\description{
Function to convert VariantAnnotation CollapsedVCF/ExpandedVCF
objects to a data frame
}
\keyword{internal}
