% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sumstats.R
\name{write_sumstats}
\alias{write_sumstats}
\title{Write sum stats file to disk}
\source{
\href{https://github.com/Bioconductor/VariantAnnotation/issues/35}{
VariantAnnotation::writeVcf has some unexpected/silent
file renaming behavior}
}
\usage{
write_sumstats(
  sumstats_dt,
  save_path,
  sep = "\\t",
  write_vcf = FALSE,
  tabix_index = FALSE,
  nThread = 1,
  return_path = FALSE,
  save_path_check = FALSE
)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics
file for the GWAS.}

\item{save_path}{File path to save formatted data. Defaults to
\code{tempfile(fileext=".tsv.gz")}.}

\item{sep}{ The separator between columns. Defaults to the character in the set \code{[,\\t |;:]} that separates the sample of rows into the most number of lines with the same number of fields. Use \code{NULL} or \code{""} to specify no separator; i.e. each line a single character column like \code{base::readLines} does.}

\item{write_vcf}{Whether to write as VCF (TRUE) or tabular file (FALSE).}

\item{tabix_index}{Index the formatted summary statistics with
\href{http://www.htslib.org/doc/tabix.html}{tabix} for fast querying.}

\item{nThread}{The number of threads to use. Experiment to see what works best for your data on your hardware.}

\item{return_path}{Return \code{save_path}.
This will have been modified in some cases
(e.g. after compressing and tabix-indexing a
previously un-compressed file).}

\item{save_path_check}{Ensure path name is valid (given the other arguments)
before writing (default: FALSE).}
}
\value{
If \code{return_path=TRUE}, returns \code{save_path}.
Else returns \code{NULL}.
}
\description{
Write sum stats file to disk
}
\examples{
path <- system.file("extdata", "eduAttainOkbay.txt",
    package = "MungeSumstats"
)
eduAttainOkbay <- read_sumstats(path = path)
write_sumstats(
    sumstats_dt = eduAttainOkbay,
    save_path = tempfile(fileext = ".tsv.gz")
)
}
