% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sumstats.R
\name{write_sumstats}
\alias{write_sumstats}
\title{Write sum stats file to disk}
\usage{
write_sumstats(
  sumstats_dt,
  save_path,
  sep,
  write_vcf = FALSE,
  tabix_index = FALSE,
  nThread = 1
)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS.}

\item{save_path}{File path to save formatted data. Defaults to \code{tempfile(fileext=".tsv.gz")}.}

\item{sep}{ The separator between columns. Defaults to the character in the set \code{[,\\t |;:]} that separates the sample of rows into the most number of lines with the same number of fields. Use \code{NULL} or \code{""} to specify no separator; i.e. each line a single character column like \code{base::readLines} does.}

\item{write_vcf}{Whether to write as VCF (TRUE) or tabular file (FALSE).}

\item{tabix_index}{Index the formatted summary statistics with \href{http://www.htslib.org/doc/tabix.html}{tabix} for fast querying.}

\item{nThread}{The number of threads to use. Experiment to see what works best for your data on your hardware.}
}
\value{
\code{VRanges} object
}
\description{
Write sum stats file to disk
}
\examples{
 eduAttainOkbay <- MungeSumstats::read_sumstats(path = system.file("extdata","eduAttainOkbay.txt",
                                                                   package="MungeSumstats"))
MungeSumstats::write_sumstats(sumstats_dt=eduAttainOkbay, 
                              save_path=tempfile(fileext=".tsv.gz"), 
                              sep="\t")
}
