% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_sumstats.R
\name{write_sumstats}
\alias{write_sumstats}
\title{Write sum stats file to disk}
\usage{
write_sumstats(
  sumstats_dt,
  save_path,
  sep,
  write_vcf = FALSE,
  tabix_index = FALSE,
  nThread = 1
)
}
\arguments{
\item{sumstats_dt}{data table obj of the summary statistics file for the GWAS.}

\item{write_vcf}{Whether to write as VCF (TRUE) or tabular file (FALSE).}

\item{check_save_out}{Output of \code{check_save_path()}.}
}
\value{
\code{VRanges} object
}
\description{
Write sum stats file to disk
}
\examples{
 eduAttainOkbay <- MungeSumstats::read_sumstats(path = system.file("extdata","eduAttainOkbay.txt",
                                                                   package="MungeSumstats"))
MungeSumstats::write_sumstats(sumstats_dt=eduAttainOkbay, 
                              save_path=tempfile(fileext=".tsv.gz"), 
                              sep="\t")
}
